Make it more likely that the patch happens
The patch only works if patchviasocket.py AND cloud-early-config are
connected to the socket at the same time. This removes most of the
long sleeps and tries more often.

Instead of trying 5 times in 31 seconds (1+2+4+8+16) we now try it
120 times in 60 seconds. The short sleeps make it more likely that
both sides are connected.

It's temporarily, ulti the Qemu Guest Agent fix is in.