Fix KVM Maintenance mode
Maintenace mode never worked, because some IDs were mixed up resulting in adding the possible hosts to the Avoid Set. You see the method is called with a number of parameters. The final one is filled like 'pool.getId()' but then in the method it compares to the hostId instead of the poolId. With this change it works fine in the bubble.