Do not serve DefaultGateway option for non-default nics
Otherwise a VM will have two default gateways and that leads to confusion in most cases.

On non-VPC routers this already worked. The option to specify the dhcpops.txt file wasn't picked-up when specified in cloud.conf. When I moved it to dnsmasq.conf (like non-VPC routers) it started working.