Fix for filling up the buffer for p.wait()
By using p.wait(), when the output of a command has a large amount of stdout or stderr, this will hang forever and cause upstream issues (reset of a VPN connection will stop working). By adding a buffer, the process will put the output into the variable and p.wait() will only check if the command finished successfully. Later, p.communicate() is used to actually gather logs. Tested on live routerVM and confirmed it fixed the issue from the GUI.

This only occurs when there are a large number of VPNs being used.